# feat: add option SWIN_torch_compile to accelerate SwinIR upscale