add option SWIN_torch_compile to accelerate SwinIR upscale torch.compile()

## Description

SwinIR (or SwinIR v2) is a SOTA upscaler, which is my preferred choice personally. But is slower on inference since it's a transformer model. However, torch.compile in PyTorch 2.0 can significantly accelerate it.

### Enhance
This PR introduces a new optition "SWIN_torch_compile" in Settings/Upscaling (disabled by default, only shows up on PyTorch 2.0 or above, and not on Windows). When enabled, it utilizes `torch.compile()` to optimize the SwinIR model. It will significant improved the inference speed of the SwinIR upscaler, by more than 2x on my RTX A4000 (card similar to a low TDP 3070ti with more VRAM).

### Limitation
The `torch.compile()` not works on Windows or PyTorch versions before 2.0, the option will not be displayed in these cases.

To "compile" the model, it will takes longer time on the first inference. Since the serialization API of the compiled model is not stable yet in PyTorch, the model needs to be recompiled every time when webui is restarted or some setting is changed. It will takes about 1~2min on my RTX A4000. Hence, the benefits of compilation may not be significant if you only need to upscale a small number of images (e.g., less than 10).

Also, the compiled model will keep in VRAM, introduced the VRAM overhead about 50 ~ 100MiB, which is rather small, and can be released (if necessary) by disable this option and performs an upscale again.

## Screenshots/videos:

The output image appears the same regardless of whether the "SWIN_torch_compile" option is enabled.

Benchmarks:
```
# tested on RTX A4000 (ECC off), CPU: AMD 5700G, no overclock

# SwinIR v1, image: 1792 x 896, swinir tile: 192, overlap: 8
SwinIR tiles: 100%|█████████████████████| 50/50 [00:21<00:00,  2.31it/s]    # baseline (without SWIN_torch_compile)
SwinIR tiles: 100%|█████████████████████| 50/50 [02:00<00:00,  2.40s/it]    # now enable SWIN_torch_compile, first run will take some times to 'compile' the model
SwinIR tiles: 100%|█████████████████████| 50/50 [00:10<00:00,  4.62it/s]    # later runs, the speed doubled
SwinIR tiles: 100%|█████████████████████| 50/50 [00:10<00:00,  4.65it/s]
SwinIR tiles: 100%|█████████████████████| 50/50 [00:10<00:00,  4.67it/s]
SwinIR tiles: 100%|█████████████████████| 50/50 [00:10<00:00,  4.65it/s]
SwinIR tiles: 100%|█████████████████████| 50/50 [00:10<00:00,  4.66it/s]

# SwinIR v2, same config
SwinIR tiles: 100%|█████████████████████| 50/50 [00:14<00:00,  3.41it/s]    # baseline (without SWIN_torch_compile)
SwinIR tiles: 100%|█████████████████████| 50/50 [00:14<00:00,  3.45it/s]
SwinIR tiles: 100%|█████████████████████| 50/50 [01:26<00:00,  1.72s/it]    # now enable SWIN_torch_compile, first run will take some times to 'compile' the model
SwinIR tiles: 100%|█████████████████████| 50/50 [00:05<00:00,  8.60it/s]    # later runs, the speed is more than doubled
SwinIR tiles: 100%|█████████████████████| 50/50 [00:05<00:00,  8.60it/s]
SwinIR tiles: 100%|█████████████████████| 50/50 [00:05<00:00,  8.59it/s]
SwinIR tiles: 100%|█████████████████████| 50/50 [00:05<00:00,  8.51it/s]
SwinIR tiles: 100%|█████████████████████| 50/50 [00:05<00:00,  8.57it/s]
```

## Checklist:

- [x] I have read [contributing wiki page](https://github.com/AUTOMATIC1111/stable-diffusion-webui/wiki/Contributing)
- [x] I have performed a self-review of my own code
- [x] My code follows the [style guidelines](https://github.com/AUTOMATIC1111/stable-diffusion-webui/wiki/Contributing#code-style)
- [x] My code passes [tests](https://github.com/AUTOMATIC1111/stable-diffusion-webui/wiki/Tests)